[PR] 견적 관리 구현

+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/entity/order/OrderProduct.java
@@ -18,5 +18,5 @@ public class OrderProduct {
     @ManyToOne
     @JoinColumn(name = "order_code")
     private Order order;
-    private Integer productCode;
+    private Long productCode;
 }
